Finished analogue axes, fixed a few other problems which may have been fixed somewhere else. Perhaps a branch went missing?

+++ src/lib/net/connection.hpp
@@ -44,7 +44,7 @@ namespace droidpad {
 
 	class DPData {
 		public:
-			std::vector<uint16_t> axes;
+			std::vector<int> axes;
 			std::vector<bool> buttons;
 	};
